"pull master"

+++ routes/listing.js
@@ -7,7 +7,11 @@ const userData = data.users;
 
 router.get('/', (req, res) => {
 	try {
-		res.render('pages/listing', { title: 'Post Listing' });
+		if (req.session && req.session.user) {
+			res.render('pages/listing', { title: 'New Listing' });
+		} else {
+			res.redirect('/login');
+		}
 	} catch (e) {
 		res.status(400).render('pages/error', { errorMessage: 'listing page Error', title: 'Error' });
 	}
@@ -26,12 +30,11 @@ router.post('/', async (req, res) => {
 				req.session.user,
 				req.body.itemName,
 				req.body.itemDesc,
-				req.body.price,
 				addr[0],
 				addr[1],
 				addr[2],
 				req.body.postal_code,
-				req.body.price
+				req.body.price,
 			);
 
 			let user = await userData.get(req.session.user);
@@ -58,18 +61,28 @@ router.post('/', async (req, res) => {
 
 router.get('/images/add/:id', async (req, res) => {
 	try {
-		let turls = await itemData.getThumbnailURL(String(req.params.id));
-		res.render('pages/images', { images: turls });
+		if (req.session && req.session.user) {
+			let turls = await itemData.getThumbnailURL(String(req.params.id));
+			res.render('pages/images', { images: turls });
+		} else {
+			res.redirect('/login');
+		}
 	} catch (e) {
 		console.log(e);
 		res.status(400).render('pages/error', { errorMessage: 'listing page Error', title: 'Error' });
 	}
 });
 
+
+
 router.post('/images/add/addCloudinary/:id', (req, res) => {
 	try {
-		itemData.updateCloudinary(String(req.params.id), req.body.url, req.body.turl);
-		res.sendStatus(200);
+		if (req.session && req.session.user) {
+			itemData.updateCloudinary(String(req.params.id), req.body.url, req.body.turl);
+			res.sendStatus(200);
+		} else {
+			res.sendStatus(400);
+		}
 	} catch (e) {
 		console.log(e);
 		res.sendStatus(400);
@@ -78,8 +91,12 @@ router.post('/images/add/addCloudinary/:id', (req, res) => {
 
 router.post('/edit/addCloudinary/:id', (req, res) => {
 	try {
-		itemData.updateCloudinary(String(req.params.id), req.body.url, req.body.turl);
-		res.sendStatus(200);
+		if (req.session && req.session.user) {
+			itemData.updateCloudinary(String(req.params.id), req.body.url, req.body.turl);
+			res.sendStatus(200);
+		} else {
+			res.sendStatus(400);
+		}
 	} catch (e) {
 		console.log(e);
 		res.sendStatus(400);
@@ -88,13 +105,17 @@ router.post('/edit/addCloudinary/:id', (req, res) => {
 
 router.post('/edit/removeCloudinary/:id', (req, res) => {
 	try {
-		itemData.removeCloudinary(String(req.params.id), req.body.url, req.body.turl);
-		cloudinary.api.delete_resources([req.body.publicId],
-		  function(error, result){
-				console.log(result);
-			}
-		);
-		return;
+		if (req.session && req.session.user) {
+			itemData.removeCloudinary(String(req.params.id), req.body.url, req.body.turl);
+			cloudinary.api.delete_resources([req.body.publicId],
+			  function(error, result){
+					console.log(result);
+				}
+			);
+			res.sendStatus(200);
+		} else {
+			res.sendStatus(400);
+		}
 	} catch (e) {
 		console.log(e);
 		res.status(400).render('pages/error', { errorMessage: 'listing page Error', title: 'Error' });
@@ -103,30 +124,63 @@ router.post('/edit/removeCloudinary/:id', (req, res) => {
 
 router.get('/edit/:id', async (req, res) => {
 	try {
-		let item = await itemData.get(String(req.params.id));
-		let urls = await itemData.getCloudinaryURL(String(req.params.id));
-		let turls = await itemData.getThumbnailURL(String(req.params.id));
+		if (req.session && req.session.user) {
+			let item = await itemData.get(String(req.params.id));
+			let urls = await itemData.getCloudinaryURL(String(req.params.id));
+			let turls = await itemData.getThumbnailURL(String(req.params.id));
 
-		let images = [];
+			let images = [];
+			let toggleStatus = "";
 
-		for (index in urls) {
-			images.push({
-				url: urls[index],
-				turl: turls[index]
-			});
+			for (index in urls) {
+				images.push({
+					url: urls[index],
+					turl: turls[index]
+				});
+			}
+
+			if (item.status === "open") {
+				toggleStatus = "Close Listing"
+			} else if (item.status === "close") {
+				toggleStatus = "Open Listing"
+			} else {
+				toggleStatus = "NANI?";
+			}
+
+			res.render('pages/viewListing', { item: item, images: images, status: toggleStatus });
+		} else {
+			res.redirect('/login');
 		}
-		console.log(images);
-		res.render('pages/viewListing', { item: item, images: images });
 	} catch (e) {
 		console.log(e);
 		res.status(400).render('pages/error', { errorMessage: 'listing page Error', title: 'Error' });
 	}
 });
 
-router.put('/listing', async (req, res) => {
+router.post('/update/:id', async (req, res) => {
 	try {
-		let item = await itemData.get(req.params.id);
-		res.render('pages/viewListing', { item: item });
+		if (req.session && req.session.user) {
+			// let item = await itemData.get(req.params.id);
+			let addr = req.body.address.split(',');
+			let item = await itemData.update(
+				req.params.id,
+				req.body.startDate,
+				req.body.endDate,
+				req.body.reqOrPost,
+				req.body.status,
+				req.session.user,
+				req.body.itemName,
+				req.body.itemDesc,
+				addr[0],
+				addr[1],
+				addr[2],
+				req.body.postal_code,
+				req.body.price,
+			);
+			res.redirect('/profile');
+		} else {
+			res.redirect('/login');
+		}
 	} catch (e) {
 		console.log(e);
 		res.status(400).render('pages/error', { errorMessage: 'listing page Error', title: 'Error' });
